+++ src/common/core/UUID.ts
@@ -15,12 +15,14 @@ export class UUID {
 
     constructor() {
         let d = new Date().getTime();
-        if (window.performance && typeof window.performance.now === "function") {
+        if (window.performance && typeof window.performance.now === 'function') {
             d += performance.now();
         }
         this.uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
-            let r = (d + Math.random() * 16) % 16 | 0;
+            // tslint:disable-next-line:no-bitwise
+            const r = (d + Math.random() * 16) % 16 | 0;
             d = Math.floor(d / 16);
+            // tslint:disable-next-line:no-bitwise
             return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16);
         });
     }