+++ src/common/core/graphics/rect.ts
@@ -24,22 +24,27 @@ export interface IRectOverlapInfos {
 }
 
 export class Rect {
+    public left: number;
+    public top: number;
+    public width: number;
+    public height: number;
+
     public static equals(r1: Rect, r2: Rect) {
         return r1 && r2 && r1.left === r2.left && r1.top === r2.top && r1.width === r2.width && r1.height === r2.height;
     }
 
     public static union(r1: Rect, r2: Rect) {
-        return Rect.fromLTRB(Math.min(r1.left, r2.left), Math.min(r1.top, r2.top), Math.max(r1.right(), r2.right()), Math.max(r1.bottom(), r2.bottom()));
+        return Rect.fromLTRB(Math.min(r1.left, r2.left), Math.min(r1.top, r2.top), Math.max(r1.left + r1.width, r2.left + r2.width), Math.max(r1.top + r1.height, r2.top + r2.height));
     }
 
     public static overlapInfos(rect1: Rect, rect2: Rect): IRectOverlapInfos {
-        let x = Math.max(0, Math.min(rect1.right(), rect2.right()) - Math.max(rect1.left, rect2.left));
-        let y = Math.max(0, Math.min(rect1.bottom(), rect2.bottom()) - Math.max(rect1.top, rect2.top));
+        const x = Math.max(0, Math.min(rect1.right, rect2.right) - Math.max(rect1.left, rect2.left));
+        const y = Math.max(0, Math.min(rect1.bottom, rect2.bottom) - Math.max(rect1.top, rect2.top));
         return {
             area: x * y,
-            direction: x > y ? RectOverlapDirection.horizontal : RectOverlapDirection.vertical,
-            height: y,
             width: x,
+            height: y,
+            direction: x > y ? RectOverlapDirection.horizontal : RectOverlapDirection.vertical,
         };
     }
 
@@ -51,14 +56,9 @@ export class Rect {
         return Rect.fromLTRB(Math.min(p1.left, p2.left), Math.min(p1.top, p2.top), Math.max(p1.left, p2.left), Math.max(p1.top, p2.top));
     }
 
-    public left: number;
-    public top: number;
-    public width: number;
-    public height: number;
-
     constructor(left?: number | Object, top?: number, width?: number, height?: number) {
         if (typeof left === 'object') {
-            let bounds = left as any;
+            const bounds = left as any;
             this.left = bounds.left || 0;
             this.top = bounds.top || 0;
             this.width = bounds.width || 0;
@@ -71,11 +71,19 @@ export class Rect {
         }
     }
 
-    public right() {
+    public set right(value: number) {
+        this.width = value - this.left;
+    }
+
+    public get right() {
         return this.left + this.width;
     }
 
-    public bottom() {
+    public set bottom(value: number) {
+        this.height = value - this.top;
+    }
+
+    public get bottom() {
         return this.top + this.height;
     }
 
@@ -89,29 +97,29 @@ export class Rect {
     }
 
     public adjacent(bounds: Rect) {
-        return bounds.left <= this.right() &&
-            bounds.right() >= this.left &&
-            bounds.top <= this.bottom() &&
-            bounds.bottom() >= this.top;
+        return bounds.left <= this.right &&
+            bounds.right >= this.left &&
+            bounds.top <= this.bottom &&
+            bounds.bottom >= this.top;
     }
 
     public contains(bounds: Rect) {
-        return bounds.left >= this.left && bounds.right() <= this.right() && bounds.top >= this.top && bounds.bottom() <= this.bottom();
+        return bounds.left >= this.left && bounds.right <= this.right && bounds.top >= this.top && bounds.bottom <= this.bottom;
     }
 
     public containsPoint(point: Position) {
-        return point.left >= this.left && point.left <= this.right() && point.top >= this.top && point.top <= this.bottom();
+        return point.left >= this.left && point.left <= this.right && point.top >= this.top && point.top <= this.bottom;
     }
 
     public intersectWith(bounds: Rect) {
-        return bounds.left < this.right() &&
-            bounds.right() > this.left &&
-            bounds.top < this.bottom() &&
-            bounds.bottom() > this.top;
+        return bounds.left < this.right &&
+            bounds.right > this.left &&
+            bounds.top < this.bottom &&
+            bounds.bottom > this.top;
     }
 
-    public union(bounds: Rect) {
-        return new Rect(Math.min(bounds.left, this.left), Math.min(bounds.top, this.top), Math.max(bounds.right(), this.right()), Math.max(bounds.bottom(), this.bottom()));
+    public isEmpty() {
+        return !this.width || !this.height;
     }
 
     public clone() {