+++ src/common/core/graphics/size.ts
@@ -10,12 +10,12 @@
  */
 
 export class Size {
+    public width: number;
+    public height: number;
+
     public static equals(s1: Size, s2: Size) {
         return s1.width === s2.width && s1.height === s2.height;
     }
-
-    public width: number;
-    public height: number;
 
     constructor(width?: number, height?: number) {
         this.width = width || 0;