+++ src/component/monaco-editor/options/editor-options.model.ts
@@ -9,9 +9,9 @@
  *
  */
 
-import {IEditorLanguage} from "./editor-language.model";
-import {IEditorScrollbarOptions} from "./editor-scrollbar-options";
-import {IEditorTheme} from "./editor-theme.component";
+import {IEditorLanguage} from './editor-language.model';
+import {IEditorScrollbarOptions} from './editor-scrollbar-options';
+import {IEditorTheme} from './editor-theme.component';
 /**
  * Configuration options for the editor.
  */